Move FontTag to the generic module
